Read datalog interactively with $EDITOR
`biscuit generate`, `biscuit attenuate` (and `biscuit verify`, optionally),
read datalog provided by the user. In addition to inline datalog, and datalog read from
a user-provided file (or stdin, for `biscuit generate`), this commit adds the possibility
of reading datalog from a temporary file created by the CLI, with $EDITOR opened to edit it,
exactly like with `git commit`.

Error management is crude, and EDITOR is defaulted to `vim` if not present, but it works.

The temporary file is given a `.datalog` extension so that editor has a chance to enable
language support (eg syntax coloring). I'm not super happy with the extension, as the
datalog used by biscuit is not standard (because of its extensions, and of the use of
`<-` instead of the more common `:-` sigil for rules. Bikeshed away!