Fail with an explicit error message when using both $EDITOR and stdin
Reading from stdin before spawning $EDITOR made the text editor run
with an already used stdin, which made it do strange things when stdin
was provided data through a pipe (interestingly, not when the data was
provided interactively through the terminal).

This is a fringe case, so it's better to fail explicitly.