feat: support JSON output in `biscuit inspect`
The `--json` flag on `biscuit inspect` now provides results in a structured JSON format:

```json
{
  "token": {
    "sealed": false,
    "root_key_id": null,
    "blocks": [
      {
        "code": "user(\"1234\");\n",
        "external_key": null,
        "revocation_id": "0ca759b390fbbb20461d9a2ca0b335371020262cbefc50854ee5cdc000339a877788f40baa0eb4af005eab0f26e556c0a86f063618d720c0619d7a67320de707"
      },
      {
        "code": "user(\"4567\");\n",
        "external_key": null,
        "revocation_id": "c7e85b0c3bad81a1f381985042df1bd6b04f5c0029b0c38ec8565d767474953e9087af66ac396dd246f416038b7c02d2e943a1c5a5a4f282b7b3f4fcbc19fb0d"
      }
    ]
  },
  "signatures_check": true,
  "auth": {
    "policies": [
      "allow if user($u)"
    ],
    "result": {
      "Ok": [
        0,
        "allow if user($u)"
      ]
    }
  },
  "query": {
    "query": "user($u) <- user($u)",
    "query_all": true,
    "facts": [
      "user(\"1234\")",
      "user(\"4567\")"
    ]
  }
}
```

If either the signature, the authorization, or the query fail during execution, a JSON value is still sent to stdout, with structured errors, an error message is sent to stderr, and the command exits with a non-zero error code.

Evaluation still aborts for input errors (missing files, invalid datalog input for the authorizer, etc). In that case, nothing is sent to stdout, an error message is sent to stderr, and the command exits with a non-zero error code.

This change required extensive modifications: the implementation interleaved console output with computations. This meant that even if an error was raised and execution aborted, text output would still be shown. For JSON output, this doesn't work, as all results need to be collected before being displayed.

The inspection function now returns an `InspectionResults` value, that can either be rendered as text, or as JSON. To be able to generate this value, even in the case of authorization errors, the execution logic had to be updated to not abort early in case of:

- failed signature checks
- failed authorization
- failed queries