inspect: allow dumping and loading policies snapshot
A policies snapshot is built from an `AuthorizerBuilder`, and does not contain a biscuit. It allows to easily share authorizer contents that can be added to a biscuit during verification.

Name is not set in stone. I don’t think using `AuthorizerBuilder` in the CLI would make a lot of sense, since it’s really tied to the rust implementation